requirements.txt
discord.py
flask
python-dotenv
flask_cors
datetime